#37 NeedToAttendType 추가
