+++ attendance-ios/Source/Model/Models.swift
@@ -20,14 +20,14 @@ struct Session: Codable {
     let title: String
     let date: String // yyyy-mm-dd hh:mm:ss
     let description: String
-    let isOff: Bool
+    let type: NeedToAttendType
 
     enum CodingKeys: String, CodingKey {
         case sessionId = "session_id"
         case title
         case date
         case description
-        case isOff = "is_off"
+        case type
     }
 }
 
@@ -50,40 +50,4 @@ enum TeamType: String, Codable {
     case ios = "IOS"
     case web = "WEB"
     case allRounder = "ALL_ROUNDER"
-}
-
-struct Attendance: Codable {
-    let sesstionId: Int
-    let attendanceType: AttendanceType
-}
-
-enum AttendanceType: Int, Codable {
-    case absence
-    case tardy
-    case attendance
-    case attendanceMarked
-
-    var text: String {
-        switch self {
-        case .absence: return "결석"
-        case .tardy: return "지각"
-        case .attendance: return "출석"
-        case .attendanceMarked: return "출석 인정"
-        }
-    }
-
-    var point: Int {
-        switch self {
-        case .absence: return -10
-        case .tardy: return -5
-        case .attendance: return 0
-        case .attendanceMarked: return 0
-        }
-    }
-}
-
-enum NeedToAttendType: String {
-    case needAttendance = "NEED_ATTENDANCE"
-    case dontNeedAttendance = "DONT_NEED_ATTENDANCE"
-    case dayOff = "DAY_OFF"
 }
